+++ template/src/sign.cpp
@@ -39,8 +39,8 @@ Hitbox Sign::get_hitbox() const
 	vec2 position = mc.position;
 	position.x -= width / 2;
 	position.y += width / 2;
-	Square top(position, width);
-	Square bot(add(position, { 0.f, width }), width);
+	Square top(position, (int)width);
+	Square bot(add(position, { 0.f, width }), (int)width);
 	squares[0] = top;
 	squares[1] = bot;
